+++ src/main/java/com/recruitment/placement_system/repository/ApplicationRepository.java
@@ -8,7 +8,7 @@ import com.recruitment.placement_system.entity.Application;
 
 public interface ApplicationRepository extends JpaRepository<Application, Integer> {
 
-    // ✅ NEW: Find all applications for a specific drive
-    List<Application> findByDriveId(int driveId);
+    // ✅ Find all applications for a specific drive
+    List<Application> findByDriveId(Long driveId);
 
 }
